requirements.txt
numpy
pandas
nba_api
plotly
scipy
streamlit-autorefresh
Pillow
matplotlib
retry
opencv-python
scikit-learn
ultralytics
tqdm
jsonlines
moviepy
tensorflow[and-cuda]
stqdm
streamlit-javascript
pytz
gekko
flask
retry
pymc
arviz
pygwalker
pypokerengine